[BEAM-8343] Moved BeamTable classes to a more appropriate location
